test: move mock-serializer tests into their own file
snapshot.test.ts carries a pre-existing TTY-dependent inline snapshot
('error snapshots') that fails whenever stderr is not a TTY. Keep that
file untouched and put the mock serializer tests alongside it in
snapshot-tests/mock-serializer.test.ts.
